requirements.txt
whisperx
torch
torchaudio
ffmpeg-python
